move `magic.py` into scriptrunner
